Recover method ownership, lifetimes, and prototypes

+++ decomp/ST.exe/functions/00428B20/decomp.c
@@ -115,7 +115,7 @@ LAB_00428c07:
   }
   iVar2 = (local_14 - local_10) * uVar4;
   iVar3 = Library::DKW::LIB::FUN_006aac70(iVar2);
-  FUN_006b0290(iVar3,param_4[2],0,0,(int)param_1,param_2,*param_4,param_4[1],param_4[2],param_4[3]);
+  FUN_006b0290(iVar3,param_4[2],0,0,param_1,param_2,*param_4,param_4[1],param_4[2],param_4[3]);
   DAT_007f4dd0 = DAT_007f4dd0 + iVar2;
   return iVar3;
 }
