+++ decomp/ST.exe/functions/006C8910/decomp.c
@@ -8,7 +8,7 @@ void FUN_006c8910(uint *param_1,int *param_2)
   if (param_2 != (int *)0x0) {
     *param_2 = 0;
   }
-  iVar1 = FUN_006c87f0((int)param_1);
+  iVar1 = FUN_006c87f0(param_1);
   puVar2 = (uint *)Library::DKW::LIB::FUN_006aac70(iVar1);
   if (puVar2 == (uint *)0x0) {
     return;
